+++ elasticdl_client/common/k8s_client.py
@@ -23,7 +23,7 @@ from kubernetes.client import V1EnvVar, V1EnvVarSource, V1ObjectFieldSelector
 from elasticdl_client.common.k8s_resource import parse as parse_resource
 from elasticdl_client.common.k8s_volume import parse_volume_and_mount
 from elasticdl_client.common.log_utils import default_logger as logger
-from elasticdl_client.common.model_utils import load_module
+from elasticdl_client.common.module_utils import load_module
 
 ELASTICDL_APP_NAME = "elasticdl"
 ELASTICDL_JOB_KEY = "elasticdl-job-name"
